Simplify editor context and form state handling
