+++ react/components/EditorContainer/Sidebar/BlockEditor/utils.test.ts
@@ -1,6 +1,4 @@
-import EXTENSIONS from '../__fixtures__/extensions'
-
-import { getIsDefaultContent, getIsSitewide } from './utils'
+import { getIsDefaultContent } from './utils'
 
 describe('getIsDefaultContent', () => {
   it('should return true when origin is declared (comes from app)', () => {
@@ -9,31 +7,5 @@ describe('getIsDefaultContent', () => {
 
   it('should return false when origin is null (declared by user)', () => {
     expect(getIsDefaultContent({ origin: null })).toBe(false)
-  })
-})
-
-describe('getIsSitewide', () => {
-  it('should return true for AFTER', () => {
-    const mockEditTreePath = 'store.home/$after_footer'
-
-    expect(getIsSitewide(EXTENSIONS, mockEditTreePath)).toBe(true)
-  })
-
-  it('should return true for AROUND', () => {
-    const mockEditTreePath = 'store.home/$around_homeWrapper'
-
-    expect(getIsSitewide(EXTENSIONS, mockEditTreePath)).toBe(true)
-  })
-
-  it('should return true for BEFORE', () => {
-    const mockEditTreePath = 'store.home/$before_header.full'
-
-    expect(getIsSitewide(EXTENSIONS, mockEditTreePath)).toBe(true)
-  })
-
-  it('should return false for blocks without role', () => {
-    const mockEditTreePath = 'store.home/carousel#home'
-
-    expect(getIsSitewide(EXTENSIONS, mockEditTreePath)).toBe(false)
   })
 })
