+++ react/utils/components/index.ts
@@ -127,7 +127,7 @@ export const getExtension = (
     composition = 'blocks',
     configurationsIds = [],
     content = {},
-    contentId = '',
+    contentIds = [],
     contentMapId = '',
     hasContentSchema = false,
     implementationIndex = 0,
@@ -150,7 +150,7 @@ export const getExtension = (
     composition,
     configurationsIds,
     content,
-    contentId,
+    contentIds,
     contentMapId,
     hasContentSchema,
     implementationIndex,
